requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0

# Text processing
nltk>=3.6.0
textblob>=0.15.3

# Machine learning (for future real model integration)
scikit-learn>=1.0.0
transformers>=4.15.0
torch>=1.10.0

# Custom model training and observability
shap>=0.41.0
requests>=2.28.0

# Data handling
json5>=0.9.0
pyyaml>=6.0

# Model frameworks
transformers>=4.15.0
torch>=1.10.0
nemo-toolkit>=1.20.0

# NeMo specific dependencies
nemo[asr]>=1.20.0
librosa>=0.9.0
soundfile>=0.12.0

# Video processing dependencies
opencv-python>=4.5.0
ffmpeg-python>=0.2.0

# Experiment tracking
wandb>=0.15.0

# Logging and monitoring
loguru>=0.6.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development tools
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Web framework for Zoom integration
flask>=2.3.0
python-dotenv>=1.0.0

# Optional: For production deployment
gunicorn>=20.1.0
fastapi>=0.70.0
uvicorn>=0.15.0 